primeiro desafio e parte do segundo concluido

+++ semana7/labenusers/src/Components/ListaUsuarios.js
@@ -2,6 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import axios from 'axios';
 import Formulario from './Formulario'
+import Detalhes from './Detalhes'
 
 const ContainerLista = styled.div`
 margin: 0 auto;
@@ -21,6 +22,8 @@ export default class ListaUsuarios extends React.Component {
 
     state = {
         usuarios: [],
+        renderizaPagina: "lista",
+        usuarioId: ""
 
     }
 
@@ -37,42 +40,60 @@ export default class ListaUsuarios extends React.Component {
             }
             )
             .then((res) => {
-                this.setState({usuarios: res.data});
+                this.setState({ usuarios: res.data });
             })
-            
+
     }
 
     deleteUsuario = (idUsuario) => {
-        axios.delete(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${idUsuario}`,
-        {
-            headers:{
-                Authorization: "guilherme-mota-cruz"
-            }
-        })
-        .then((res) => {
-            this.getUsuarios()
-            alert("O usuário foi excluído!")
-            
-        })
-        .catch((err) => {
-            alert("Erro ao remover o usuário")
-        })
+        if (window.confirm("Tem certeza de que deseja deletar o usuário?")) {
+            axios.delete(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${idUsuario}`,
+                {
+                    headers: {
+                        Authorization: "guilherme-mota-cruz"
+                    }
+                })
+                .then((res) => {
+                    this.getUsuarios()
+                    alert("O usuário foi excluído!")
+
+                })
+                .catch((err) => {
+                    alert("Erro ao remover o usuário")
+                })
+        }
+    }
+
+    paginaDetalhes = (id) => {
+        if (this.state.renderizaPagina === "lista") {
+            this.setState({ renderizaPagina: "detalhes", usuarioId: id });
+        } else {
+            this.setState({ renderizaPagina: "lista", usuarioId: "" });
+        }
     }
 
     render() {
-        const listaDeUsuario = this.state.usuarios.map((usuario) =>(
+        const listaDeUsuario = this.state.usuarios.map((usuario) => (
             <BlocoNome>
-            <p key={usuario.id}>{usuario.name}</p>
-            <button onClick={() => this.deleteUsuario(usuario.id)}>excluir</button>
+                <p onClick={() => this.paginaDetalhes(usuario.id)} key={usuario.id}>{usuario.name}</p>
+                <button onClick={() => this.deleteUsuario(usuario.id)}>excluir</button>
             </BlocoNome>
         ))
         return (
             <ContainerLista>
+
                 
-                <h2>Lista de usuários cadastrados</h2>
-                
-                <p>{listaDeUsuario}</p>
-                
+                {this.state.renderizaPagina === "lista" ? 
+                (<div>
+                    <h2>Lista de usuários cadastrados</h2>
+                    <p>{listaDeUsuario}</p>
+                </div>) 
+                : 
+                (<Detalhes
+                    mudarPagina={this.paginaDetalhes}
+                    
+                />)}
+
             </ContainerLista>
         );
     }
